Tag v1.3

+++ test/lib/OAuth2/Storage/Bootstrap.php
@@ -9,6 +9,7 @@ class Bootstrap
     private $sqlite;
     private $mongo;
     private $redis;
+    private $cassandra;
     private $configDir;
 
     public function __construct()
@@ -49,13 +50,31 @@ class Bootstrap
         if (!$this->redis) {
             if (class_exists('Predis\Client')) {
                 $redis = new \Predis\Client();
-                $redis->flushdb();
-                $this->redis = new Redis($redis);
-                $this->createRedisDb($this->redis);
+                if ($this->testRedisConnection($redis)) {
+                    $redis->flushdb();
+                    $this->redis = new Redis($redis);
+                    $this->createRedisDb($this->redis);
+                } else {
+                    $this->redis = new NullStorage('Redis', 'Unable to connect to redis server on port 6379');
+                }
+            } else {
+                $this->redis = new NullStorage('Redis', 'Missing redis library. Please run "composer.phar require predis/predis:dev-master"');
             }
         }
 
         return $this->redis;
+    }
+
+    private function testRedisConnection(\Predis\Client $redis)
+    {
+        try {
+            $redis->connect();
+        } catch (\Predis\CommunicationException $exception) {
+            // we were unable to connect to the redis server
+            return false;
+        }
+
+        return true;
     }
 
     public function getMysqlPdo()
@@ -77,16 +96,111 @@ class Bootstrap
         if (!$this->mongo) {
             $skipMongo = isset($_SERVER['SKIP_MONGO_TESTS']) && $_SERVER['SKIP_MONGO_TESTS'];
             if (!$skipMongo && class_exists('MongoClient')) {
-                $m = new \MongoClient();
-                $db = $m->oauth2_server_php;
-                $this->removeMongoDb($db);
-                $this->createMongoDb($db);
+                $mongo = new \MongoClient('mongodb://localhost:27017', array('connect' => false));
+                if ($this->testMongoConnection($mongo)) {
+                    $db = $mongo->oauth2_server_php;
+                    $this->removeMongoDb($db);
+                    $this->createMongoDb($db);
 
-                $this->mongo = new Mongo($db);
+                    $this->mongo = new Mongo($db);
+                } else {
+                    $this->mongo = new NullStorage('Mongo', 'Unable to connect to mongo server on "localhost:27017"');
+                }
+            } else {
+                $this->mongo = new NullStorage('Mongo', 'Missing mongo php extension. Please install mongo.so');
             }
         }
 
         return $this->mongo;
+    }
+
+    private function testMongoConnection(\MongoClient $mongo)
+    {
+        try {
+            $mongo->connect();
+        } catch (\MongoConnectionException $e) {
+            return false;
+        }
+
+        return true;
+    }
+
+    public function getCassandraStorage()
+    {
+        if (!$this->cassandra) {
+            if (class_exists('phpcassa\ColumnFamily')) {
+                $cassandra = new \phpcassa\Connection\ConnectionPool('oauth2_test', array('127.0.0.1:9160'));
+                if ($this->testCassandraConnection($cassandra)) {
+                    $this->removeCassandraDb();
+                    $this->cassandra = new Cassandra($cassandra);
+                    $this->createCassandraDb($this->cassandra);
+                } else {
+                    $this->cassandra = new NullStorage('Cassandra', 'Unable to connect to cassandra server on "127.0.0.1:9160"');
+                }
+            } else {
+                $this->cassandra = new NullStorage('Cassandra', 'Missing cassandra library. Please run "composer.phar require thobbs/phpcassa:dev-master"');
+            }
+        }
+
+        return $this->cassandra;
+    }
+
+    private function testCassandraConnection(\phpcassa\Connection\ConnectionPool $cassandra)
+    {
+        try {
+            new \phpcassa\SystemManager('localhost:9160');
+        } catch (\Exception $e) {
+            return false;
+        }
+
+        return true;
+    }
+
+    private function removeCassandraDb()
+    {
+        $sys = new \phpcassa\SystemManager('localhost:9160');
+
+        try {
+            $sys->drop_keyspace('oauth2_test');
+        } catch (\cassandra\InvalidRequestException $e) {
+
+        }
+    }
+
+    private function createCassandraDb(Cassandra $storage)
+    {
+        // create the cassandra keyspace and column family
+        $sys = new \phpcassa\SystemManager('localhost:9160');
+
+        $sys->create_keyspace('oauth2_test', array(
+            "strategy_class" => \phpcassa\Schema\StrategyClass::SIMPLE_STRATEGY,
+            "strategy_options" => array('replication_factor' => '1')
+        ));
+
+        $sys->create_column_family('oauth2_test', 'auth');
+
+        // populate the data
+        $storage->setClientDetails("oauth_test_client", "testpass", "http://example.com", 'implicit password');
+        $storage->setAccessToken("testtoken", "Some Client", '', time() + 1000);
+        $storage->setAuthorizationCode("testcode", "Some Client", '', '', time() + 1000);
+        $storage->setUser("testuser", "password");
+
+        $storage->setScope('supportedscope1 supportedscope2 supportedscope3 supportedscope4');
+        $storage->setScope('defaultscope1 defaultscope2', null, 'default');
+
+        $storage->setScope('clientscope1 clientscope2', 'Test Client ID');
+        $storage->setScope('clientscope1 clientscope2', 'Test Client ID', 'default');
+
+        $storage->setScope('clientscope1 clientscope2 clientscope3', 'Test Client ID 2');
+        $storage->setScope('clientscope1 clientscope2', 'Test Client ID 2', 'default');
+
+        $storage->setScope('clientscope1 clientscope2', 'Test Default Scope Client ID');
+        $storage->setScope('clientscope1 clientscope2', 'Test Default Scope Client ID', 'default');
+
+        $storage->setScope('clientscope1 clientscope2 clientscope3', 'Test Default Scope Client ID 2');
+        $storage->setScope('clientscope3', 'Test Default Scope Client ID 2', 'default');
+
+        $storage->setClientKey('oauth_test_client', $this->getTestPublicKey(), 'test_subject');
     }
 
     private function createSqliteDb(\PDO $pdo)
@@ -110,31 +224,28 @@ class Bootstrap
 
     public function runPdoSql(\PDO $pdo)
     {
-        $pdo->exec('CREATE TABLE oauth_clients (client_id TEXT, client_secret TEXT, redirect_uri TEXT, grant_types TEXT, user_id TEXT)');
+        $pdo->exec('CREATE TABLE oauth_clients (client_id TEXT, client_secret TEXT, redirect_uri TEXT, grant_types TEXT, scope TEXT, user_id TEXT, public_key TEXT)');
         $pdo->exec('CREATE TABLE oauth_access_tokens (access_token TEXT, client_id TEXT, user_id TEXT, expires DATETIME, scope TEXT)');
         $pdo->exec('CREATE TABLE oauth_authorization_codes (authorization_code TEXT, client_id TEXT, user_id TEXT, redirect_uri TEXT, expires DATETIME, scope TEXT)');
-        $pdo->exec('CREATE TABLE oauth_users (username TEXT, password TEXT, first_name TEXT, last_name TEXT)');
+        $pdo->exec('CREATE TABLE oauth_users (username TEXT, password TEXT, first_name TEXT, last_name TEXT, scope TEXT)');
         $pdo->exec('CREATE TABLE oauth_refresh_tokens (refresh_token TEXT, client_id TEXT, user_id TEXT, expires DATETIME, scope TEXT)');
-        $pdo->exec('CREATE TABLE oauth_scopes (type TEXT, scope TEXT, client_id TEXT)');
+        $pdo->exec('CREATE TABLE oauth_scopes (scope TEXT, is_default BOOLEAN)');
         $pdo->exec('CREATE TABLE oauth_public_keys (client_id TEXT, public_key TEXT, private_key TEXT, encryption_algorithm VARCHAR(100) DEFAULT "RS256")');
+        $pdo->exec('CREATE TABLE oauth_jwt (client_id VARCHAR(80), subject VARCHAR(80), public_key VARCHAR(2000))');
 
         // set up scopes
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope) VALUES ("supported", "supportedscope1 supportedscope2 supportedscope3 supportedscope4")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope) VALUES ("default", "defaultscope1 defaultscope2")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope, client_id) VALUES ("supported", "clientscope1 clientscope2", "Test Client ID")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope, client_id) VALUES ("default", "clientscope1 clientscope2", "Test Client ID")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope, client_id) VALUES ("supported", "clientscope1 clientscope2 clientscope3", "Test Client ID 2")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope, client_id) VALUES ("default", "clientscope1 clientscope2", "Test Client ID 2")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope, client_id) VALUES ("supported", "clientscope1 clientscope2", "Test Default Scope Client ID")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope, client_id) VALUES ("default", "clientscope1 clientscope2", "Test Default Scope Client ID")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope, client_id) VALUES ("supported", "clientscope1 clientscope2 clientscope3", "Test Default Scope Client ID 2")');
-        $pdo->exec('INSERT INTO oauth_scopes (type, scope, client_id) VALUES ("default", "clientscope3", "Test Default Scope Client ID 2")');
+        foreach (explode(' ', 'supportedscope1 supportedscope2 supportedscope3 supportedscope4 clientscope1 clientscope2 clientscope3') as $supportedScope) {
+            $pdo->exec(sprintf('INSERT INTO oauth_scopes (scope) VALUES ("%s")', $supportedScope));
+        }
+
+        foreach (array('defaultscope1', 'defaultscope2') as $defaultScope) {
+            $pdo->exec(sprintf('INSERT INTO oauth_scopes (scope, is_default) VALUES ("%s", 1)', $defaultScope));
+        }
 
         // set up clients
-        $pdo->exec('INSERT INTO oauth_clients (client_id, client_secret) VALUES ("Test Client ID", "TestSecret")');
-        $pdo->exec('INSERT INTO oauth_clients (client_id, client_secret) VALUES ("Test Client ID 2", "TestSecret")');
-        $pdo->exec('INSERT INTO oauth_clients (client_id, client_secret) VALUES ("Test Default Scope Client ID", "TestSecret")');
-        $pdo->exec('INSERT INTO oauth_clients (client_id, client_secret) VALUES ("Test Default Scope Client ID 2", "TestSecret")');
+        $pdo->exec('INSERT INTO oauth_clients (client_id, client_secret, scope) VALUES ("Test Client ID", "TestSecret", "clientscope1 clientscope2")');
+        $pdo->exec('INSERT INTO oauth_clients (client_id, client_secret, scope) VALUES ("Test Client ID 2", "TestSecret", "clientscope1 clientscope2 clientscope3")');
+        $pdo->exec('INSERT INTO oauth_clients (client_id, client_secret, scope) VALUES ("Test Default Scope Client ID", "TestSecret", "clientscope1 clientscope2")');
         $pdo->exec('INSERT INTO oauth_clients (client_id, client_secret, grant_types) VALUES ("oauth_test_client", "testpass", "implicit password")');
 
         // set up misc
@@ -143,7 +254,8 @@ class Bootstrap
         $pdo->exec('INSERT INTO oauth_users (username, password) VALUES ("testuser", "password")');
         $pdo->exec('INSERT INTO oauth_public_keys (client_id, public_key, private_key, encryption_algorithm) VALUES ("ClientID_One", "client_1_public", "client_1_private", "RS256")');
         $pdo->exec('INSERT INTO oauth_public_keys (client_id, public_key, private_key, encryption_algorithm) VALUES ("ClientID_Two", "client_2_public", "client_2_private", "RS256")');
-        $pdo->exec(sprintf('INSERT INTO oauth_public_keys (client_id, public_key, private_key, encryption_algorithm) VALUES (NULL, "%s", "%s", "RS256")', file_get_contents($this->configDir.'/keys/id_rsa.pub'), file_get_contents($this->configDir.'/keys/id_rsa')));
+        $pdo->exec(sprintf('INSERT INTO oauth_public_keys (client_id, public_key, private_key, encryption_algorithm) VALUES (NULL, "%s", "%s", "RS256")', $this->getTestPublicKey(), $this->getTestPrivateKey()));
+        $pdo->exec(sprintf('INSERT INTO oauth_jwt (client_id, subject, public_key) VALUES ("oauth_test_client", "test_subject", "%s")', $this->getTestPublicKey()));
     }
 
     public function removeMysqlDb(\PDO $pdo)
@@ -163,10 +275,33 @@ class Bootstrap
 
     private function createMongoDb(\MongoDB $db)
     {
-        $db->oauth_clients->insert(array('client_id' => "oauth_test_client", 'client_secret' => "testpass", 'redirect_uri' => "http://example.com", 'grant_types' => 'implicit password'));
-        $db->oauth_access_tokens->insert(array('access_token' => "testtoken", 'client_id' => "Some Client"));
-        $db->oauth_authorization_codes->insert(array('authorization_code' => "testcode", 'client_id' => "Some Client"));
-        $db->oauth_users->insert(array('username' => "testuser", 'password' => "password"));
+        $db->oauth_clients->insert(array(
+            'client_id' => "oauth_test_client",
+            'client_secret' => "testpass",
+            'redirect_uri' => "http://example.com",
+            'grant_types' => 'implicit password'
+        ));
+
+        $db->oauth_access_tokens->insert(array(
+            'access_token' => "testtoken",
+            'client_id' => "Some Client"
+        ));
+
+        $db->oauth_authorization_codes->insert(array(
+            'authorization_code' => "testcode",
+            'client_id' => "Some Client"
+        ));
+
+        $db->oauth_users->insert(array(
+            'username' => "testuser",
+            'password' => "password"
+        ));
+
+        $db->oauth_jwt->insert(array(
+            'client_id' => 'oauth_test_client',
+            'key'       => $this->getTestPublicKey(),
+            'subject'   => 'test_subject',
+        ));
     }
 
     private function createRedisDb(Redis $storage)
@@ -190,10 +325,22 @@ class Bootstrap
 
         $storage->setScope('clientscope1 clientscope2 clientscope3', 'Test Default Scope Client ID 2');
         $storage->setScope('clientscope3', 'Test Default Scope Client ID 2', 'default');
+
+        $storage->setClientKey('oauth_test_client', $this->getTestPublicKey(), 'test_subject');
     }
 
     public function removeMongoDb(\MongoDB $db)
     {
         $db->drop();
+    }
+
+    public function getTestPublicKey()
+    {
+        return file_get_contents(__DIR__.'/../../../config/keys/id_rsa.pub');
+    }
+
+    private function getTestPrivateKey()
+    {
+        return file_get_contents(__DIR__.'/../../../config/keys/id_rsa');
     }
 }
